requirements.txt
fastapi 
uvicorn 
pymongo 
passlib[bcrypt] 
python-jose[cryptography] 
python-dotenv 
authlib